pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "checkngn"
version = "0.1.3"
description = "A lightweight Python DSL for setting up business intelligence rules that can be configured without code."
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    {name = "AhnafCodes"}
]
urls = {Homepage = "https://github.com/AhnafCodes/checkngn"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=2.3.3",
    "numpy>=2.4.0",
    "python-dateutil",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "mock>=5.0.0",
    "pytest>=8.0.0",
    "pandas>=2.3.3"
]

[tool.hatch.build.targets.wheel]
packages = ["checkngn"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]
